SQL playgrounds: dedup CSV/Parquet helpers and schema settings stores
Three near-identical copies of the RFC-4180 CSV parser, the
table-name-from-filename helper, and the parquet-wasm loader lived in
the SQLite hook, the DuckDB import module, and the Postgres import
module. Consolidate them into sql/utils/importUtils.ts and re-export
from the dialect modules so call sites keep working.

The DuckDB and Postgres settings stores were also byte-for-byte
identical except for the export and type names. Move the body to a
new sql/stores/createSchemaSettingsStore factory and have each
dialect call it. The SQLite store is left in place — its theme
defaults and PRAGMA-tab integration differ.

Net: ~220 fewer lines, single source of truth for CSV import and the
schema-aware settings shape.
